refactor: 노티피케이션 리팩토링, sse 커스텀훅으로, 타입 파일로 저장
