+++ src/types/NotiList.d.ts
@@ -1,0 +1,14 @@
+export interface NotiList {
+  id: number;
+  notification_type: string;
+  actor: number;
+  content: string;
+  related_id: number;
+  created_at: string;
+  is_read: boolean;
+  redirect_url: {
+    redirect_url: string;
+    method: string;
+    token_required: string;
+  };
+}
